overhauled tests, allow binary_search on immutable slices
bad definition of the OrdSubsetSliceExt trait did not allow
immutable slices to be binaryly searched. This was fixed.
(Note: this was also broken before one of the earlier commits that
       deleted one of the impls, as the impl for [T] did not affect &[T])

extended test suite to cover all functions save for .*_by()
variants (those are used by all the others anyway) and checks
trait implementations for arrays, vecs, slices and references