Reduce OrdSubsetSliceExt impls to only slices
object safety rules are going to make the current implementation
illegal.
Because this trait is only there to add methods to slices, it's fitting
to reduce it to only doing that rather than doing as_ref() or as_mut()
conversions implicitly. Derefs and UnsizedCoercions will keep the methods
callable on most types. The rest has to convert explicitly to slices.